Refactor Login to use thunk actions
