--- src/redux/user/userActions.js
@@ -1,3 +1,0 @@
-import { setUser, resetUser } from './userReducer';
-
-export { setUser, resetUser };
